각도값 int->float

+++ app/src/main/java/kr/hs/emirim/s2125/mirim_project_0726_1/MainActivity.java
@@ -46,7 +46,7 @@ public class MainActivity extends AppCompatActivity {
                 item.setChecked(true);
                 return true;
             case R.id.btn:
-                imgV.setRotation(Integer.parseInt(edit.getText().toString()));
+                imgV.setRotation(Float.parseFloat(edit.getText().toString()));
                 return true;
 
         }
